+++ packages/valibot-form-data/src/index.ts
@@ -1,6 +1,9 @@
 export { checkbox } from "./lib/checkbox.ts";
 export { file } from "./lib/file.ts";
-export { formData, preprocessFormData } from "./lib/form-data.ts";
+export { formData } from "./lib/form-data.ts";
 export { numeric } from "./lib/numeric.ts";
 export { repeatable, repeatableOfType } from "./lib/repeatable.ts";
 export { text } from "./lib/text.ts";
+
+// Re-export from form-utils for convenience
+export { preprocessFormData } from "@typed-web/form-utils";